Fixing errors

+++ navigate-to.php
@@ -117,25 +117,27 @@ require_once $_SERVER['DOCUMENT_ROOT'].'/climate-canary/includes/header.php';
                         </a>';
             } else {
                 foreach( $data as $row => $field ) {
-					echo '  <div class="row divided">
-                                <div class="col col-xs-2">
-                                    <a href="favorite.php?type=add&id='.base64_encode($field['id']).'"><i class="material-icons">star_border</i></a>
-                                </div>
-                
-                                <a href="navigation-confirm.php?from=fAddress&fromLat=fLat&fromLng=fLng&to='.$field['address'].'&toLat='.$field['latitude'].'&toLng='.$field['longitude'].'" class="col col-xs-7 address">
-                                    '.( !empty( $field['name'] ) ? $field['name'] : $field['address'] ).'
-                                </a>
-                                
-                                <div class="col col-xs-3">
-                                    <!--<div class="col col-xs-6">
-                                        <a href="address-edit.php?id='.base64_encode($field['id']).'"><i class="material-icons">edit</i></a>
-                                    </div>-->
-                                    
-                                    <div class="col col-xs-6">
-                                        <a href="address-delete.php?id='.base64_encode($field['id']).'" onClick="return confirm(\'Weet je zeker dat je dit adres wilt verwijderen?\')"><i class="material-icons">delete</i></a>
+                    if( $field['favorite'] == 0 ) {
+						echo '  <div class="row divided">
+                                    <div class="col col-xs-2">
+                                        <a href="favorite.php?type=add&id='.base64_encode( $field['id'] ).'"><i class="material-icons">star_border</i></a>
                                     </div>
-                                </div>
-                            </div>';
+                    
+                                    <a href="navigation-confirm.php?from=fAddress&fromLat=fLat&fromLng=fLng&to='.$field['address'].'&toLat='.$field['latitude'].'&toLng='.$field['longitude'].'" class="col col-xs-7 address">
+                                        '.( !empty( $field['name'] ) ? $field['name'] : $field['address'] ).'
+                                    </a>
+                                    
+                                    <div class="col col-xs-3">
+                                        <!--<div class="col col-xs-6">
+                                            <a href="address-edit.php?id='.base64_encode( $field['id'] ).'"><i class="material-icons">edit</i></a>
+                                        </div>-->
+                                        
+                                        <div class="col col-xs-6">
+                                            <a href="address-delete.php?id='.base64_encode( $field['id'] ).'" onClick="return confirm(\'Weet je zeker dat je dit adres wilt verwijderen?\')"><i class="material-icons">delete</i></a>
+                                        </div>
+                                    </div>
+                                </div>';
+					}
                 }
             }
             ?>
